requirements.txt
openai
pyautogui
keyboard
opencv-python
pillow
python-dotenv
mss
